navbar updated
